Implemented state management via Agents
