+++ .gitignore
@@ -5,3 +5,4 @@ erl_crash.dump
 *.ez
 .DS_Store
 /*/.DS_Store
+/priv
